PortfolioDetailView.

+++ src/InvestLens.Model/SecurityInfo.cs
@@ -13,6 +13,6 @@ public class SecurityInfo
     public int Count { get; init; }
     public double AveragePrice { get; init; }
     public double CurrentPrice { get; init; }
-    public decimal TotalPrice { get; init; }
-    public decimal Profit { get; init; }
+    public double TotalPrice { get; init; }
+    public double Profit { get; init; }
 }
